+++ odb-examples/hello/driver.cxx
@@ -22,7 +22,7 @@ main (int argc, char* argv[])
   {
     unique_ptr<database> db (create_database (argc, argv));
 
-    unsigned long john_id, joe_id;
+    unsigned long long john_id, joe_id;
 
     // Create a few persistent person objects.
     //
